refactor: 重构部分api代码 更新doc文档

+++ FMSystem/FMSystem/ViewModels/LessonViewModel.cs
@@ -10,8 +10,8 @@ namespace FMSystem.ViewModels
         public int SectionId { get; set; }
         public int LessonNo { get; set; }
         public int CoachId { get; set; }
-        public int StartDate { get; set; }
-        public int EndDate { get; set; }
+        public long StartDate { get; set; }
+        public long EndDate { get; set; }
         public string State { get; set; }
     }
 }
